Add files via upload
Those are some programs in which I'm working on those days. The one named Brent is actually some code that I got from the link that is in the archive, the code performs Brent's method to find the root of a function. From now on, I'm gonna  try to implement this method on the Pain VI Arb. Further I will build something that does de same kind of root calculation to the Continued Fraction problem, also using ArbNumerics. Wish me luck!